# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bank_fineract_models.ErrorResponse:
    properties:
      errorMessage:
        type: string
      errorStackTrace:
        type: string
      errorStatusCode:
        type: integer
    type: object
...
  bank_fineract_models.Payment:
    properties:
      accountNumber:
        type: string
      accountRef:
        type: string
      amount:
        type: number
      approver:
        type: string
      attachment:
        items:
          type: integer
        type: array
      auditedOn:
        type: string
      bankReference:
        type: string
      barcode:
        items:
          type: integer
        type: array
      channelReference:
        type: string
      channelRequestType:
        type: string
      checker:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      deletedOn:
        type: string
      description:
        type: string
      enabled:
        type: integer
      id:
        type: integer
      locale:
        type: string
      maker:
        type: string
      msisdn:
        type: string
      name:
        type: string
      paymentNumber:
        type: integer
      qrcode:
        items:
          type: integer
        type: array
      stage:
        type: string
      state:
        type: string
      status:
        type: string
      syncToken:
        type: integer
      timestamp:
        type: integer
      transactionReference:
        type: string
      transactionType:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      version:
        type: integer
      workflowLevels:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
host: localhost:9001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a finaract middleware API  documentation .
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fineract Middleware API
  version: "1.0"
paths:
...
  /middleware/api/v1/payment/id/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve Payment by by PaymentID
      operationId: Retrieve Payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bank_fineract_models.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bank_fineract_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/bank_fineract_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bank_fineract_models.ErrorResponse'
      summary: Retrieve Payment by PaymentID
      tags:
      - Payment
...
  /middleware/api/v1/payment/update/{id}:
    patch:
      consumes:
      - application/json
      description: get string by ID
      operationId: update-payment-by-id-int
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/bank_fineract_models.Payment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bank_fineract_models.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/bank_fineract_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/bank_fineract_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/bank_fineract_models.ErrorResponse'
      summary: Update an Payment
      tags:
      - Payment
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"